backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.36.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5
pydantic-settings==2.11.0
PyJWT==2.10.1
httpx==0.28.1
python-dotenv==1.1.1
pgvector==0.4.1
sentence-transformers==5.1.2
apscheduler==3.11.0
greenlet==3.2.4
psycopg2-binary==2.9.10
numpy==1.26.4
